Added actions and reducers for wishlist functionality
