--- src/actions/actionTypes.js
@@ -1,2 +1,0 @@
-export const INCREMENT = 'INCREMENT';
-export const DECREMENT = 'DECREMENT';
