+++ src/reducers/index.js
@@ -1,25 +1,8 @@
 import { combineReducers } from 'redux';
-import { INCREMENT, DECREMENT } from '../actions/actionTypes';
-
-const initialState = {
-  count: 0,
-};
-
-function reducer(state = initialState, action) {
-  switch (action.type) {
-  case INCREMENT:
-    return {
-      count: state.count + 1,
-    };
-  case DECREMENT:
-    return {
-      count: state.count - 1,
-    };
-  default:
-    return state;
-  }
-}
+import reducerCounter from './reducerCounter';
+import reducerWishlist from './reducerWishlist';
 
 export default combineReducers({
-  reducer,
+  reducerCounter,
+  reducerWishlist,
 });
